gpt/gpt_requirements.txt
openai
python-dotenv
pypdf
opencv-python
pytesseract
pdf2image
ppt2pdf
